fix: restructure Layout to rander PostLayout/ShowcaseLayout when type is specified accordingly
