--- src/components/news/Layout.tsx
@@ -1,52 +1,0 @@
-import Header from "@/components/Header";
-import NavBar from "@/components/NavBar";
-import TopLines from "@/components/TopLines";
-import Footer from "../homepage/footer";
-
-type Props = {
-  page_header?: string;
-  children: React.ReactNode;
-};
-export default function Layout({ page_header, children }: Props) {
-  return (
-    <>
-      <Header title={"News"} />
-      <NavBar />
-      <TopLines />
-      <div className="flex flex-col">
-        <div className="self-center flex w-full max-w-[1068px] flex-col px-5 max-md:max-w-full mt-[100px]">
-          <h1 className="font-fredoka">{page_header}</h1>
-          <div className="self-center w-full mt-10 max-md:max-w-full max-md:mt-10">
-            <div className="gap-5 flex max-md:flex-col max-md:items-stretch max-md:gap-0">
-              <main>{children}</main>
-            </div>
-          </div>
-        </div>
-      </div>
-      <Footer />
-      <style jsx>
-        {`
-          .root {
-            display: block;
-            padding: 4rem 0;
-            box-sizing: border-box;
-            height: 100%;
-          }
-          main {
-            display: flex;
-            min-height: 100%;
-          }
-          @media (min-width: 769px) {
-            .root {
-              display: flex;
-              flex: 1 0 auto;
-            }
-            main {
-              flex: 1 0 auto;
-            }
-          }
-        `}
-      </style>
-    </>
-  );
-}
