+++ src/components/showcase/ShowcaseLayout.tsx
@@ -2,21 +2,15 @@ import React from "react";
 import Link from "next/link";
 import Image from "next/image";
 import styles from "@/public/styles/posts.module.css";
-import Author from "../news/Author";
 import Copyright from "../news/Copyright";
 import Date from "../news/Date";
-import Layout from "../news/Layout";
 import BasicMeta from "../news/meta/BasicMeta";
-import JsonLdMeta from "../news/meta/JsonLdMeta";
 import OpenGraphMeta from "../news/meta/OpenGraphMeta";
 import TwitterCardMeta from "../news/meta/TwitterCardMeta";
-import TagButton from "../news/TagButton";
 import { getFellow } from "../../lib/people";
 import { getTag } from "../../lib/tags";
-import { MDXRemote, MDXRemoteSerializeResult } from "next-mdx-remote";
-import Footer from "../homepage/footer";
 
-type Props = {
+export type ShowcaseLayoutProps = {
   title: string;
   date: Date;
   slug: string;
@@ -38,12 +32,12 @@ export default function ShowcaseLayout({
   tags,
   description = "",
   children,
-}: Props) {
+}: ShowcaseLayoutProps) {
   const keywords = tags ? tags.map((it) => getTag(it).name) : [];
   // const authorName = getAuthor(author).name;
   const fellow = getFellow(fellowName);
   return (
-    <Layout>
+    <>
       <BasicMeta
         url={`/showcase/${slug}`}
         title={title}
@@ -104,146 +98,6 @@ export default function ShowcaseLayout({
           <Copyright />
         </footer>
       </div>
-      <style jsx>
-        {`
-          .metadata div {
-            display: inline-block;
-            margin-right: 0.5rem;
-          }
-          article {
-            flex: 1 0 auto;
-          }
-          h1 {
-            margin: 0 0 0.5rem;
-            font-size: 2.3rem;
-          }
-          .backlink {
-            color: grey;
-            margin-bottom: 1em;
-          }
-          .tag-list {
-            list-style: none;
-            text-align: right;
-            margin: 1.75rem 0 0 0;
-            padding: 0;
-          }
-          .tag-list li {
-            display: inline-block;
-            margin-left: 0.5rem;
-          }
-          .social-list {
-            margin-top: 3rem;
-            text-align: center;
-          }
-
-          @media (min-width: 769px) {
-            .container {
-              display: flex;
-              flex-direction: column;
-            }
-          }
-        `}
-      </style>
-      <style global jsx>
-        {`
-          /* Syntax highlighting */
-          .token.comment,
-          .token.prolog,
-          .token.doctype,
-          .token.cdata,
-          .token.plain-text {
-            color: #6a737d;
-          }
-
-          .token.atrule,
-          .token.attr-value,
-          .token.keyword,
-          .token.operator {
-            color: #d73a49;
-          }
-
-          .token.property,
-          .token.tag,
-          .token.boolean,
-          .token.number,
-          .token.constant,
-          .token.symbol,
-          .token.deleted {
-            color: #22863a;
-          }
-
-          .token.selector,
-          .token.attr-name,
-          .token.string,
-          .token.char,
-          .token.builtin,
-          .token.inserted {
-            color: #032f62;
-          }
-
-          .token.function,
-          .token.class-name {
-            color: #6f42c1;
-          }
-
-          /* language-specific */
-
-          /* JSX */
-          .language-jsx .token.punctuation,
-          .language-jsx .token.tag .token.punctuation,
-          .language-jsx .token.tag .token.script,
-          .language-jsx .token.plain-text {
-            color: #24292e;
-          }
-
-          .language-jsx .token.tag .token.attr-name {
-            color: #6f42c1;
-          }
-
-          .language-jsx .token.tag .token.class-name {
-            color: #005cc5;
-          }
-
-          .language-jsx .token.tag .token.script-punctuation,
-          .language-jsx .token.attr-value .token.punctuation:first-child {
-            color: #d73a49;
-          }
-
-          .language-jsx .token.attr-value {
-            color: #032f62;
-          }
-
-          .language-jsx span[class="comment"] {
-            color: pink;
-          }
-
-          /* HTML */
-          .language-html .token.tag .token.punctuation {
-            color: #24292e;
-          }
-
-          .language-html .token.tag .token.attr-name {
-            color: #6f42c1;
-          }
-
-          .language-html .token.tag .token.attr-value,
-          .language-html
-            .token.tag
-            .token.attr-value
-            .token.punctuation:not(:first-child) {
-            color: #032f62;
-          }
-
-          /* CSS */
-          .language-css .token.selector {
-            color: #6f42c1;
-          }
-
-          .language-css .token.property {
-            color: #005cc5;
-          }
-        `}
-      </style>
-    </Layout>
+    </>
   );
 }
